@@ -1,7 +1,7 @@
 from flask import Flask, jsonify
 from flask_cors import CORS
 from models import db, Article
-from config import Config  # Ensure this line is correct
+from config import Config
 
 app = Flask(__name__)
 app.config.from_object(Config)
@@ -10,8 +10,13 @@ CORS(app)
 
 @app.route('/news', methods=['GET'])
 def get_news():
-    articles = Article.query.order_by(Article.published.desc()).limit(10).all()
-    return jsonify([article.to_dict() for article in articles])
+    try:
+        articles = Article.query.order_by(Article.published.desc()).limit(10).all()
+        print(f"Fetched articles: {articles}")  
+        return jsonify([article.to_dict() for article in articles])
+    except Exception as e:
+        print(f"Error fetching articles: {e}")
+        return jsonify({"error": "Error fetching articles"}), 500
 
 if __name__ == '__main__':
     with app.app_context():
