@@ -6,14 +6,14 @@ class Article(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.String(255), nullable=False)
     link = db.Column(db.String(255), nullable=False)
-    published = db.Column(db.String(100), nullable=False)
-    source = db.Column(db.String(255), nullable=False)
+    published = db.Column(db.DateTime, nullable=False)
+    source = db.Column(db.String(50), nullable=False)
 
     def to_dict(self):
         return {
             'id': self.id,
             'title': self.title,
             'link': self.link,
-            'published': self.published,
+            'published': self.published.isoformat(),
             'source': self.source
         }
