@@ -1,4 +1,3 @@
-// src/App.js
 import React from 'react';
 import './App.css';
 import NewsList from './components/NewsList';
@@ -6,7 +5,9 @@ import NewsList from './components/NewsList';
 function App() {
   return (
     <div className="App">
-      <h1>Bulgarian News</h1>
+      <header className="App-header">
+        <h1>Latest News in Bulgaria</h1>
+      </header>
       <NewsList />
     </div>
   );
